+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- N/A
+
+## [2.5.4]
+
 - Fixed bug where Sleeper leagues without divisions would fail to load
 - Fixed bug where Yahoo leagues could not load league IDs for previous years
 
@@ -230,7 +234,8 @@ All notable changes to this project will be documented in this file.
 
 ### Initial Release
 
-[Unreleased]: https://github.com/joeyagreco/leeger/compare/v2.5.3...HEAD
+[Unreleased]: https://github.com/joeyagreco/leeger/compare/v2.5.4...HEAD
+[2.5.4]: https://github.com/joeyagreco/leeger/releases/tag/v2.5.4
 [2.5.3]: https://github.com/joeyagreco/leeger/releases/tag/v2.5.3
 [2.5.0]: https://github.com/joeyagreco/leeger/releases/tag/v2.5.0
 [2.4.0]: https://github.com/joeyagreco/leeger/releases/tag/v2.4.0